c/types: moved transaction rpc types to separate file
