+++ src/v/cluster/CMakeLists.txt
@@ -229,6 +229,7 @@ v_cc_library(
     topic_recovery_validator.cc
     tm_stm_types.cc
     tx_errc.cc
+    tx_protocol_types.cc
   DEPS
     Seastar::seastar
     bootstrap_rpc
