+++ src/v/kafka/server/group.h
@@ -13,8 +13,8 @@
 #include "base/seastarx.h"
 #include "cluster/fwd.h"
 #include "cluster/simple_batch_builder.h"
+#include "cluster/tx_protocol_types.h"
 #include "cluster/tx_utils.h"
-#include "cluster/types.h"
 #include "container/fragmented_vector.h"
 #include "features/feature_table.h"
 #include "kafka/group_probe.h"
